.gemini/antigravity/scratch/healthie-backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
bcrypt
google-generativeai
python-dotenv
waitress
pillow
